Move authorization logic into new class
New class `AuthorizedClient` to seamlessly handle authorization logic in
the client. Now, we'll be able to use this class as an interface to
authorize and query the Asana client. This will make it easy to
choose whether to read authorization from a recent session (through
local storage or cookies), or to use oauth popup flow.