Move form and submit to Explorer
The form and submit button previously were in RouteEntry, but this
was mainly because of the order I developered the components. It
makes more sense for the form/button to include all input fields.
Right now, I've put them in the Explorer, but would be good to move
them to a standalone component in the future.

This also has the benefit of allowing the enter key to submit the form.