Prefer GET actions as the default for a resource
When switching resources, we previously would choose the first action path
in the list. While this was convenient, it was a bit odd since we only
support GET actions in the interactive tester. We now will choose a GET
request if one exists, and if not then will choose the first in the list.